requirements.txt
azure-storage-queue==12.1.5